[refactor] lib대신 supabaseClient로 일원화
